perlPackages: make meta.position point to the right location
Currently on master meta.position points to lib/attrsets.nix:362 for
all perlPackages. The culprit turns out to be lib.recursiveUpdate which
does the heavy lifting of merging the meta set provided in
perl-packages.nix and the preset values for homepage and platforms.
This function internally uses lib.zipAttrsWithNames.

The fix is just removing lib.recursiveUpdate and merging the meta set
manually because the rest of the stuff we are looking to merge in
buildPerlPackage is not nested.

Note that this fix only corrects the situation for perlPackages that set
meta.description: If meta.description is missing, stdenv.mkDerivation
tries to find name and version which are currently always redefined in
buildPerlPackage, so meta.position will point to
`pkgs/development/perl-modules/generic/default.nix` in those cases.
However the removal of the `name` and `version` redefinitions there are
a pending TODO, so this should fix itself at some point. Also it is
entirely possible to just add descriptions for the remaining
perlPackages without one (which wouldn't be good anyways)!