pythonPackages.afdko: fix tests for fonttools >= 4.21.0
Missed this while testing the fonttools update because my machine
couldn't handle the amount of master rebuilds at the time. fonttools
changed something in its xml output which causes afdko tests to fail
which diff against an expected output. We can simply pull in a patch
from their development branch to fix this.