wal-g: enable support for brotli compression
Brotli support was always on until v0.2.10. It is enabled by default
in the wal-g's 'official' releases and build instructions, so it makes
sense to enable it in nixpkgs too.

wal-g has a Makefile (not used by nixpkgs) which statically links to
brotli v1.0.7 (a C library). Nixpkgs dynamically links to brotli
v1.0.9.