beets: refactor external plugins into top-level package set
Instead of managing external plugins in the beets derivation, we
introduce a new top-level package set beetsExternalPlugins which the
beets derivation receives as an input. This change doesn't affect how
the beets derivation is built or overridden, so the change won't be
noticed by users, but it makes hydra evaluate and build external plugins
which should benefit users of those plugins and prevent future
regressions as we have experienced (currently on master
beets-alternatives fails to evaluate, but this wasn't picked up by
ofborg nor hydra).