lorri: 1.2.0 -> 1.3.0
We forked lorri to nix-community, so this release also moves us over
to the new official repository.

Release notes:

Fix the build loop.

Previously, any change (for example a direnv ping or a change in the
nix files) would add a new build invocation to the queue, and the
builds would all be done one after the other.

However, a new build will always use the newest state of the files
anyway, so the CPU time spent on all the other builds will be
wasted (and hog your processor).

Now lorri will only

    finish the current build (if running)
    schedule at maximum one additional build if requested

This should improve the resource use drastically in some situations.

---

Make lorri daemon exit with exit code 0 instead of 130/143 on
SIGINT or SIGTERM.

---

Add lorri self-upgrade branch sub-subcommand.
This enables us to point users to a branch name, in order to test out
fixes from repository branches.