emacs: do not use project package from elpa for emacs >= 28
The `project` package bundled with Emacs 28 is newer than the one
provided by GNU ELPA. This may or may not change in the future.

For now it is definitely the case and since the package is under heavy
development a lot of features are missing in the one on GNU ELPA which in
turn is newer than the one that comes bundled with Emacs 27.