pkgs.dockerTools.buildLayeredImage: support custom layer packing strategies
The default algorithm does not work particularly well for some use cases, such
as applications with very large dependency closures when we want to optimize
iterative releases of the same application container rather than optimize across
different application containers.

Currently there's limited sharing, since almost all of the packages are in the
final layer along with the app itself. This commit adds a new layering strategy
which will emphasize cache hits for the application itself and the dependencies
closest to the application, which are the most likely to change, while leaving
the more-likely-to-be-constant low-level dependencies smooshed into the first layer.

Since the current algorithm likely works well for many use cases, this leaves
the default unchanged, but enhances `buildLayeredImage` to be able to take a
per-nix-expression custom layering algorithm as a parameter.