julia: deprecate julia_1 alias
This is potentially controversial, but it is unclear to me whether
julia_1 should refer to either the latest stable or LTS release of 1.x.
We may want to revisit this when/if we get 2.x, but as it stands the
Julia release process [1] makes it clear that there will only ever be a
single supported stable and LTS release at any given time which should
speak in favour of using the julia-stable and julia-lts aliases instead.

[1]: https://julialang.org/blog/2019/08/release-process/#long_term_support