nixos/prometheus/exporters: fix eval if only `openFirewall = true;` is set
If `openFirewall = true`, but no `firewallFilter` is set, the evaluation
fails with the following error:

    The option `services.prometheus.exporters.node.firewallFilter` is defined both null and
    not null, in `/home/ma27/Projects/nixpkgs/nixos/modules/services/monitoring/prometheus/exporters.nix'
    and `/home/ma27/Projects/nixpkgs/nixos/modules/services/monitoring/prometheus/exporters.nix'.

Originally introduced by me in #115185. The problem is that
`mkOptionDefault` has - as its name suggests - the same priority as the
default-value of the option `firewallFilter` and thus it conflicts if
this declaration and the actual default value are set which is the case
if `firewallFilter` isn't specified somewhere else.