libtiff: fix two security issues
CVE-2020-35523:
An integer overflow flaw was found in libtiff that exists in the
tif_getimage.c file. This flaw allows an attacker to inject and execute
arbitrary code when a user opens a crafted TIFF file. The highest
threat from this vulnerability is to confidentiality, integrity, as
well as system availability.

CVE-2020-35524:
A heap-based buffer overflow flaw was found in libtiff in the handling
of TIFF images in libtiff's TIFF2PDF tool. A specially crafted TIFF
file can lead to arbitrary code execution. The highest threat from this
vulnerability is to confidentiality, integrity, as well as system
availability.

Fixes: CVE-2020-35523, CVE-2020-35524