ocamlPackages.luv: init at 0.5.7
Enabling tests is not feasible as the test suite takes > 45min (haven't
even run it to completion yet) which is not a good idea in ocamlPackages
where you often have to build from source. Additionally it would require
fetching the git repository since the test suite is not contained in the
release tarball.

Co-authored-by: locallycompact <dan.firth@homotopic.tech>