nixos/switch-to-configuration: Implement more generic decisions for config activations in containers
Actually, using `re{load,start}IfChanged` isn't the best decision for
containers because some containers have to be reloaded or restarted
depending on what has changed. For instance, a new bind-mount requires a
`machinectl reboot`, but a change in the NixOS config only needs a
`systemctl reload` (which runs `switch-to-configuration` inside the
container).

To model this, I decided to add four keywords and an option
`activation.strategy` to declarative containers:

* `strategy = "none"` means that the container will be entirely ignored
  by `switch-to-configuration`.

* `strategy = "restart"` will always `machinectl reboot` the container
  if a change was detected.

* `strategy = "reload"` will always `systemctl reload` the container if
  a change was detected.

* `strategy = "dynamic"` will check what has changed inside the
  container. If only the NixOS config inside the container has changed,
  a reload will be scheduled, otherwise a restart.

Always did a nearly full rewrite of the activation test to cover several
corner-cases and combination of such settings.