ocaml-ng.ocamlPackages*: only build whitelisted packages on hydra
recurseIntoAttrs creates _a lot_ of new build jobs for hydra for often
little gain, i. e. the older package sets have a lot of broken packages
that aren't really used by anyone, effectively wasting CPU time.
Nevertheless it is nice to have them evaluated to find serious
evaluation errors and to have the compilers and essential build tools
as well as common dependencies built by hydra.

This is achieved in this commit by conditionally adding an override
setting hydraPlatforms = [] for all except a few whitelisted packages
for the non-standard ocamlPackages sets. Currently the list is very
short, but it can be expanded easily (suggestions are welcome).

A glaring issue of this approach is that we need to constantly keep
whitelist updated if we use non-standard ocamlPackages elsewhere or even
expose it at the top level since the hydraPlatforms = [] would be
propagated.