neko: try to list applying licenses more accurately
As a side note: This change shows why `with` can be dangerous business:
It doesn't shadow any existing bindings which can be unexpected. If I
were to use with licenses; [ … ] here, zlib in the with block would
actually be the zlib passed via the function arguments instead of the
zlib from licenses which would be expected. This was what caused the
previous eval error.