rclone: build with 'cmount' support
'rclone cmount' provides the same functionality as 'rclone mount'. They
both provide FUSE integration, but 'cmount' is backed by libfuse while
'mount' is backed by a pure go reimplementation of libfuse.

'cmount' support is required for continued support for FUSE on macOS
because the traditional 'rclone mount' is no longer supported for
macOS[1]. Although 'rclone mount' would work on macOS with recent
versions of rclone, it's aliased to 'rclone cmount'.

This change is not strictly required for Linux, but it still has its
merits. According to the upstream maintainer, the libfuse implementation
is "a bit more compatible and reliable"[2] than the go implementation.
Taking this into account, I've also enabled 'cmount' support on Linux by
default.

[1]: https://github.com/rclone/rclone/issues/4393
[2]: https://github.com/restic/restic/issues/3096#issuecomment-727215997