nixos/lib/testing-python: effectively disable line length check
black imposes a very strict line length limit of 88 characters by
default and requires to break any lines longer than that. Since this
check is done on the output of code generation this has no real merit
over the actual source as the source is not a python file, but usually a
nix string in a nix expression.

Additionally we have a pretty powerful programming language to
_generate_ a test script by using string interpolation and other
capabilities we often utilize in the module system. With linting enabled
this becomes a chore since you have to e. g. worry about your lines not
getting too long without knowing in advance how long they are going to
be.