nixos/doc/manual/Makefile: Do not format
Formatting has been neglected. Running `make` would format dozens
of files, which is a great way of scaring away newcomers and those
with less git experience. It would also annoy the heck out of
regular contributors.

The purpose of formatting is to avoid a small annoyance, so it
should not become a big annoyance that makes people give up on
their work.