tanka: 0.14.0 -> 0.15.0
Update Tanka to the latest release.

Change `meta.homepage` to point to the Tanka website instead of the
GitHub repository.

Add the `subPackage` attribute and additional ldflags arguments to
`buildFlagsArray`. Both are needed to fix the issue where `tk --version`
would print `dev` instead of the version string. The ldflags arguments
were copied from the Makefile in the Tanka repository.