jamulus: clarify gpl2Plus; update github owner
license clarified according to e.g.
https://github.com/jamulussoftware/jamulus/issues/1330#issuecomment-805271880
as suggested by
https://github.com/NixOS/nixpkgs#issuecomment-804793871

the github owner was apparently changed:
https://github.com/corrados/jamulus
redirects to
https://github.com/jamulussoftware/jamulus
now.
I assume that creating this redirect requires control
of the original account; so that we can infer that the
transfer is intentional and "authorized" (and the new
repo is not some malicious malware fork or something)