linux: enable AQTION
This enables the atlantic module that provides the driver for aQuantia
AQtion AQC107/AQC108 chipset, which is an increasingly common 10G NIC
embedded in motherboards and available on various PCI cards.

Enabling this driver by default ensures that the installation experience
is smooth and doesn't involve building a new kernel on another machine.