bitwarden_rs: remove failing integration test
This test is hitting a timeout because it's trying to build
firefox-unwrapped. The idea is good, test that the web service works with
different database backends, but building firefox is way too big a task for a
small service like this.

Here's the logs from one of the failing tests:

https://hydra.nixos.org/build/140338120

From the looks of it, the only other test to use firefox-unwrapped and selenium
in the same way is nginx-etag, which also times out:

https://hydra.nixos.org/build/140337879

If there is no alternative way of automating the integration tests, I therefore
propose that we delete them. Right now, the only purpose they serve is to block
newer releases on unstable.