drone: add option to build oss edition
By default drone is distributed under an unfree license that only
permits non-commercial or evaluation use. To build the OSS edition, two
tags have to be passed as build flags. This commit adds an option to
build the OSS edition to the package and defines drone-oss as an alias
for this.