fvwm: add readline, fix man build, refactor fetch
The fvwm.1 man page was missing for the standard build of fvwm. While
digging, I also noticed that readline was not included (which, when
included, makes FvwmConsole much more pleasant to use). Based on those
two items, the following changes are in this commit.

- Add libxslt for xsltproc tool, used in building documentation

- Explicitly direct creation of man pages (prior to this change, the
  fvwm.1 man page was not built)

- Explicitly direct HTML documentation not to be built

- Add readline for use in FvwmConsole

While changing things in this package anyway, I made the following
adjustments.

- Use fetchFromGitHub instead of fetchUrl

- Reformat with nixfmt