haskellPackages.lsp-test: temporarily remove override
This override currently breaks eval, but is probably still necessary in
some form. However the change to fix it is non-trivial and not quickly
actionable as hls-plugins-api currently is broken so the intended target
of the override (haskell-language-server) can't be tested.

Something in a similar spirit would be pinning lsp-test to 0.14.0.0
which would however require to pick this patch from
haskell-language-server master:

https://github.com/haskell/haskell-language-server/commit/6d1f1a55e39a9fd905bea13d6144125fcb42a594

To me this sounds like a promising solution as it also adjusts to lsp
1.2 which we now have in haskellPackages. Supposedly it also fixes some
tests, so maybe we can remove the dontCheck.

cc @maralorn