haskellPackages: restrict a few packages depending on haskell-gi
haskell-gi and haskell-gi-base have released a 0.25.0 version on
hackage, but hackage nightly still has an 0.24.* version.

However there are some packages depending on haskell-gi like gi-vte
which are not in stackage, but have releases on hackage which address
the changes in 0.25.0. The resulting mismatched versions result in
compilation failures which are addressed by restricting these packages
to a version supporting 0.24.* temporarily until Stackage Nightly
catches up.

The most prominent package affected by this breakage was termonad which
depends gi-vte, cc @cdepillabout on this.