nixos/btrbk: init
Adding support for configuring the btrbk backup utility via
the configuration.nix. Atm this is more a less a direct mapping
from nix syntax into the configuration syntax of the tool.

The unit test requires a rather big testdisk (120M). This is
due to the fact, that btrfs images cant be smaller than that.