Revert "Merge pull request #117886 from risicle/ris-pygments-tests"
This is a pattern that should be avoided. So far hypothesis is the only
package using it if I am correct, let's not add more. Instead, we should solve the underlying issue.
That is, we build, install and test within one derivation. That we should split up.
Doing this more will only result in trouble.

This reverts commit 49e0bbb333ad0f55a8dd24d61c57e4ebde45db08, reversing
changes made to 2f2a55496a21b20ae7ad9b8b193c83b90ff8256a.