nixos/containers-next: add read-only `nixos.containers.rendered` option
This option is an attr-set that maps containers to their NixOS
configuration since `nixos.containers.instances` directly transforms the
config to a NixOS derivation. Also, the raw `nixos.containers.instances`
isn't really usable since it usually contains a list of chunks that are
evaluated by the module-system.

This is actually useful to introspect the configuration just as it's
done with e.g. `resources.machines`[1] in nixops. For instance, I'm
configuring my Prometheus scraping targets like this by gathering all active
exporters in my machines and their containers:

    { config, lib, ... }: with lib;
    let
      containers = flip mapAttrsToList machine.nixos.containers.rendered (const (x: x.config));
    in
      flip concatMap (attrValues containers)
        (c: flip concatMap (attrValues c.services.prometheus.exporters)
          (exporter:
            (optional exporter.enable "${config.networking.fqdn}:${toString exporter.port}")))

[1] https://nixos.mayflower.consulting/blog/2018/10/26/nixops-machine-configs/